+++ test/lib/index.spec.js
@@ -235,90 +235,6 @@ describe('The API', function() {
 
 describe('Mocking the file system', function() {
 
-  describe('fs.readFile(filename, [options], callback)', function() {
-    // this is provided by fs.open, fs.fstat, and fs.read
-    // so more heavily tested elsewhere
-
-    beforeEach(function() {
-      mock({
-        'path/to/file.txt': 'file content'
-      });
-    });
-    afterEach(mock.restore);
-
-    it('allows a file to be read asynchronously', function(done) {
-      fs.readFile('path/to/file.txt', function(err, data) {
-        if (err) {
-          return done(err);
-        }
-        assert.isTrue(Buffer.isBuffer(data));
-        assert.equal(String(data), 'file content');
-        done();
-      });
-    });
-
-    if (fs.promises) {
-      it('allows a file to be read asynchronously in promise', function(done) {
-        fs.promises.readFile('path/to/file.txt').then(
-          function(data) {
-            assert.isTrue(Buffer.isBuffer(data));
-            assert.equal(String(data), 'file content');
-            done();
-          },
-          function(err) {
-            done(err);
-          }
-        );
-      });
-    }
-
-    it('fails for directory', function(done) {
-      fs.readFile('path/to', function(err, data) {
-        assert.instanceOf(err, Error);
-        done();
-      });
-    });
-
-    it('fails for bad path', function(done) {
-      fs.readFile('path/to/bogus', function(err, data) {
-        assert.instanceOf(err, Error);
-        // windows has different errno for ENOENT
-        assert.equal(err.code, 'ENOENT');
-        done();
-      });
-    });
-  });
-
-  describe('fs.readFileSync(filename, [options])', function() {
-    // this is provided by fs.openSync, fs.fstatSync, and fs.readSync
-    // so more heavily tested elsewhere
-
-    beforeEach(function() {
-      mock({
-        'path/to/file.txt': 'file content'
-      });
-    });
-    afterEach(mock.restore);
-
-    it('allows a file to be read synchronously', function() {
-      const data = fs.readFileSync('path/to/file.txt');
-      assert.isTrue(Buffer.isBuffer(data));
-      assert.equal(String(data), 'file content');
-    });
-
-    it('fails for directory', function() {
-      assert.throws(function() {
-        fs.readFileSync('path/to');
-      });
-    });
-
-    it('fails for bad path', function() {
-      assert.throws(function() {
-        fs.readFileSync('path/to/bogus');
-      });
-    });
-  });
-
   describe('fs.write(fd, buffer, offset, length, position, callback)', function() {
     beforeEach(function() {
       mock({
